% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{bubble_func}
\alias{bubble_func}
\title{Bubble plot for state overlap}
\usage{
bubble_func(cell_list, state_list, state1, state2)
}
\arguments{
\item{cell_list}{vector list of cells, same length as state_list}

\item{state_list}{vector list of state, same length as state_list}

\item{state1}{string}

\item{state2}{string}
}
\value{
a shiny object of the bubble plot
}
\description{
Bubble proportionality plot for pairwise state variables.
}
