% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_inference.R
\name{chisqr_test}
\alias{chisqr_test}
\title{Pairwise Distance; Chi Square}
\usage{
chisqr_test(cat1, cat2)
}
\arguments{
\item{cat1}{vector binary list of cells in/not in cat1}

\item{cat2}{vector binary list of cells in/not in cat2}
}
\description{
Retruns a p-value for a chi square test of independence on two subsets
}
\details{
Not recommended for use, as it tests both postive and negative
correlation. For TI we usually only want to test positive correlation.
}
