% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{cluster_func}
\alias{cluster_func}
\title{One-Hot-Encode Cluster Annotations}
\usage{
cluster_func(cell_list, state_list)
}
\arguments{
\item{cell_list}{vector list of cells, same length as state_list}

\item{state_list}{vector list of state, same length as state_list}
}
\value{
dataframe A C x S dataframe of cells with one-hot-encoded states
}
\description{
One-hot-encode a metadata file for use with statorTI. In order for statorTI
to work effectively a large number of cells should belong to multiple
clusters, as opposed to in traditional one-hot-encoding.
}
