% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_inference.R
\name{hypergeo_test}
\alias{hypergeo_test}
\title{Pairwise Distance; Hypergeometric}
\usage{
hypergeo_test(cat1, cat2)
}
\arguments{
\item{cat1}{vector binary list of cells in/not in cat1}

\item{cat2}{vector binary list of cells in/not in cat2}
}
\description{
Retruns a p-value for a one-sided hypergeometric test on two subsets
}
