% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{igraph_plotter}
\alias{igraph_plotter}
\title{Plot statorTI Graphs}
\usage{
igraph_plotter(graph, graph_type, edge_labels, root)
}
\arguments{
\item{graph}{igraph object a statorTI graph to be plotted}

\item{graph_type}{string one of "mst", "cyclic", "full".}

\item{edge_labels}{bool include edge weights on the graph? Default FALSE}

\item{root}{string Only used is graph_type = "mst", set the root vertex}
}
\description{
Plot statorTI graphs using the reccomended settings.
}
