% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_inference.R
\name{jaccard_distance}
\alias{jaccard_distance}
\title{Pairwise Distance; Jaccard}
\usage{
jaccard_distance(cat1, cat2)
}
\arguments{
\item{cat1}{vector binary list of cells in/not in cat1}

\item{cat2}{vector binary list of cells in/not in cat2}
}
\description{
Retruns a distance for the jaccard distance of two sets
}
