% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{scree_plot}
\alias{scree_plot}
\title{Skree Plot for PCA}
\usage{
scree_plot(n, sdev)
}
\arguments{
\item{n}{integer number of pca components to plot}

\item{sdev}{numerical vector of standard deviations}
}
\value{
a ggplot object displaying a skree plot
}
\description{
Create a skree plot for pca.
}
