% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{statorTI}
\alias{statorTI}
\title{statorTI}
\usage{
statorTI(
  state_list,
  cell_list,
  gene_cell_matrix,
  fn = jaccard_distance,
  graph_type = "full",
  edge_labels = FALSE,
  alpha = FALSE,
  bound = FALSE,
  root = ""
)
}
\arguments{
\item{state_list}{vector list of state, same length as state_list}

\item{cell_list}{vector list of cells, same length as state_list}

\item{gene_cell_matrix}{matrix a gene cell matrix, not currently used}

\item{fn}{string the function to use to calculate the distance between
clusters on the graph}

\item{graph_type}{string one of "mst", "cyclic", "full".}

\item{edge_labels}{bool include edge weights on the graph? Default FALSE}

\item{alpha}{float the significance level for p-testing, should not be set
if bound is set}

\item{bound}{float an upper bound for pruning, should not be set if alpha
is set}

\item{root}{string Only used is graph_type = "mst", set the root vertex}
}
\description{
Perform Stator based trajectory inference
}
\details{
By default fn = jaccard_distance, a built in function. Other build in
options include hypergeo_test and chisqr_test. The later is not recommended.
Can also be a user defined function which takes numerical vectors as
input and returns a scalar.

Use bound to set the upper bound for pruning if using a metric based
function. use alpha if using a statistical test based function. Both of
these values are only used if graph_type is "cyclic".
}
